requirements.txt
numpy>=1.26.0
python-chess>=1.9.4
PyQt6>=6.6.1
pygame>=2.5.0 